extractor/requirements.txt
telethon
redis
python-dotenv